feat: Refactor and rename files
Naming file create_user_agent.go in go-useragent repository
is unnecessary duplication. It should be enough for the
file to be named create.go as the linkage to user agent
is obvious from the repo name.

Also, move UserAgent struct definition to separate file as
it is not related to either parsing or creating the UA.